+++ libs/vamp-pyin/YinUtil.cpp
@@ -28,7 +28,7 @@ YinUtil::slowDifference(const double *in, double *yinBuffer, const size_t yinBuf
     double delta ;
     int startPoint = 0;
     int endPoint = 0;
-    for (int i = 1; i < yinBufferSize; ++i) {
+    for (size_t i = 1; i < yinBufferSize; ++i) {
         yinBuffer[i] = 0;
         startPoint = yinBufferSize/2 - i/2;
         endPoint = startPoint + yinBufferSize;
